refactor : readCarNames(), readTryCount() 메서드를 InputManager 클래스로 분리
두 메서드 모두 공통적으로 입력을 다루기 때문에
Application 클래스로부터 분리
