+++ src/main/java/racingcar/Application.java
@@ -3,15 +3,14 @@ package racingcar;
 import static racingcar.Validator.validateCarNames;
 import static racingcar.Validator.validateTryCount;
 
-import camp.nextstep.edu.missionutils.Console;
 import java.util.ArrayList;
 import java.util.List;
 
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        String carNames = readCarNames();
-        String tryCount = readTryCount();
+        String carNames = InputManager.readCarNames();
+        String tryCount = InputManager.readTryCount();
 
         validateCarNames(carNames);
         validateTryCount(tryCount);
@@ -31,13 +30,4 @@ public class Application {
         return racingCarList;
     }
 
-    static String readCarNames() {
-        System.out.println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)");
-        return Console.readLine();
-    }
-
-    static String readTryCount() {
-        System.out.println("시도할 횟수는 몇 회인가요?");
-        return Console.readLine();
-    }
 }
